+++ src/main.ts
@@ -1,7 +1,10 @@
 import './assets/main.css'
 import './assets/index.css'
-
 import { createApp } from 'vue'
 import App from './App.vue'
+import router from '@/router' 
 
-createApp(App).mount('#app')
+const app = createApp(App)
+app.use(router)  
+app.mount('#app')
+
